:computer: Refactors some MarketplacesCase functions into a OfferingsCase module
