+++ apps/ex_commerce_web/test/ex_commerce_web/live/catalogue_item_live_test.exs
@@ -2,6 +2,7 @@ defmodule ExCommerceWeb.CatalogueItemLiveTest do
   @moduledoc false
 
   use ExCommerce.ContextCases.MarketplacesCase
+  use ExCommerce.ContextCases.OfferingsCase
   use ExCommerceWeb.ConnCase
 
   import Phoenix.LiveViewTest
